Move UiSpinner into new directory
