+++ src/components/animations/LoadingFallback.tsx
@@ -5,9 +5,9 @@ import type {
   ViewStyle,
 } from "react-native";
 
-import UiSpinner from "#/components/animations/UiSpinner";
 import Text from "#/components/design/Text";
 import View from "#/components/design/View";
+import UiSpinner from "#/components/ui/UiSpinner";
 
 interface LoadingFallbackProps {
   text?: string;
